pyproject.toml
[project]
name = "finpulse-pipeline"
version = "1.0.0"
description = "Financial Data Engineering Pipeline"
readme = "GEMINI.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "prefect>=2.0.0",
    "yfinance>=0.2.0",
    "python-dotenv>=1.0.0",
    "curl-cffi>=0.7.0",
    "requests>=2.30.0",
    "urllib3>=2.0.0",
    "comm>=0.2.3",
    "jupyter>=1.1.1",
    "ipykernel>=7.2.0",
    "pyarrow>=23.0.1",
    "dbt-databricks>=1.7.0",
    "elementary-data[databricks]>=0.23.2",
    "databricks-sdk>=0.20.0",
    "soda-core-spark>=3.0.0",
    "databricks-sql-connector>=3.0.0",
    "streamlit>=1.56.0",
    "plotly>=6.7.0",
    "scikit-learn>=1.7.2",
    "google-generativeai>=0.8.0",
    "groq>=1.2.0",
    "google-genai>=1.73.1",
]

[project.optional-dependencies]
# Install for Silver/Gold notebook work: uv sync --extra notebook
notebook = [
    "pyspark>=3.5.0,<4.0.0",
    "pyiceberg>=0.5.0",
]
